Move svcparse to deftree
