+++ deftree/svcparse/lexer_test.go
