+++ deftree/svcparse/scanner.go
