+++ deftree/svcparse/scanner_test.go
